+++ packages/bpk-animate-height/src/AnimateHeight.js
@@ -123,6 +123,9 @@ class AnimateHeight extends Component {
   onTransitionEnd() {
     if (this.contentElement && this.props.height === 0) {
       this.contentElement.style.display = 'none';
+      if (typeof this.props.onAnimationComplete === 'function') {
+        this.props.onAnimationComplete();
+      }
     }
   }
 
@@ -131,6 +134,7 @@ class AnimateHeight extends Component {
       children,
       duration,
       easing,
+      onAnimationComplete,
       style,
       ...rest
     } = this.props;
@@ -177,12 +181,14 @@ AnimateHeight.propTypes = {
   ]).isRequired,
   easing: PropTypes.string,
   transitionOverflow: PropTypes.string,
+  onAnimationComplete: PropTypes.func,
   style: PropTypes.object, // eslint-disable-line react/forbid-prop-types
 };
 
 AnimateHeight.defaultProps = {
   easing: 'ease',
   transitionOverflow: 'hidden',
+  onAnimationComplete: null,
   style: {},
 };